# Conflicts:
#	.gitignore

+++ modules.local/paypal-pos-php-sdk/src/Iterator/Attachment/GalleryImageIterator.php
@@ -38,7 +38,7 @@ class GalleryImageIterator implements WcProductIterator
     #[\ReturnTypeWillChange]
     public function key()
     {
-        return key($this->galleryIds);
+        return (int) key($this->galleryIds);
     }
 
     public function valid(): bool
